Extract lens URL validation helper
