+++ js/lens-url.js
@@ -1,0 +1,32 @@
+// @ts-check
+// lens-url.js - URL validation for external Knowledge Base backends.
+
+// http:// is accepted for hosts that can't leak the Bearer token across the
+// public internet: loopback, RFC1918 LAN, link-local, Tailscale CGNAT, mDNS.
+// Everything else must be https://.
+function isPrivateHost(host) {
+  if (!host) return false;
+  if (host === 'localhost') return true;
+  if (host === '::1' || host === '[::1]') return true;
+  if (host.endsWith('.local') || host.endsWith('.local.')) return true;
+  const m = /^(\d{1,3})\.(\d{1,3})\.(\d{1,3})\.(\d{1,3})$/.exec(host);
+  if (!m) return false;
+  const a = +m[1], b = +m[2], c = +m[3], d = +m[4];
+  if (a > 255 || b > 255 || c > 255 || d > 255) return false;
+  if (a === 10) return true;
+  if (a === 127) return true;
+  if (a === 169 && b === 254) return true;
+  if (a === 172 && b >= 16 && b <= 31) return true;
+  if (a === 192 && b === 168) return true;
+  if (a === 100 && b >= 64 && b <= 127) return true;
+  return false;
+}
+
+export function isValidLensUrl(url) {
+  if (typeof url !== 'string' || !url) return false;
+  let u;
+  try { u = new URL(url); } catch { return false; }
+  if (u.protocol === 'https:') return true;
+  if (u.protocol === 'http:') return isPrivateHost(u.hostname);
+  return false;
+}
